+++ CHANGES.md
@@ -13,6 +13,7 @@
  - [fleche] This achieves an almost 50% memory reduction for example
    when opening all of HoTT .v files (@ejgallego, @SkySkimmer,
    @bhaktishh, #744)
+ - [memo] Provide API to query Hashtbl stats (@ejgallego, #753)
 
 # coq-lsp 0.1.10: Hasta el 40 de Mayo _en effect_...
 ----------------------------------------------------